requirements.txt
google-generativeai
langchain-google-genai
langchain
ollama
Pillow
transformers
ml_dtypes
streamlit
uvicorn
fastapi
